backend/requirements.txt

django
djangorestframework
duckdb